move TestPlatformHelperVersioningGetStatus into domain and clean up mocks
