+++ tests/platform_helper/domain/test_versioning.py
@@ -11,6 +11,8 @@ from dbt_platform_helper.constants import PLATFORM_HELPER_VERSION_FILE
 from dbt_platform_helper.domain.versioning import PlatformHelperVersioning
 from dbt_platform_helper.domain.versioning import PlatformHelperVersionNotFoundException
 from dbt_platform_helper.domain.versioning import skip_version_check
+from dbt_platform_helper.providers.config import ConfigProvider
+from dbt_platform_helper.providers.files import FileProvider
 from dbt_platform_helper.providers.io import ClickIOProvider
 from dbt_platform_helper.providers.semantic_version import SemanticVersion
 from dbt_platform_helper.providers.version import LocalVersionProvider
@@ -27,6 +29,16 @@ def mock_local_version_provider():
 @pytest.fixture
 def mock_io_provider():
     return Mock(spec=ClickIOProvider)
+
+
+@pytest.fixture
+def mock_config_provider():
+    return Mock(spec=ConfigProvider)
+
+
+@pytest.fixture
+def mock_file_provider():
+    return Mock(spec=FileProvider)
 
 
 @pytest.fixture
@@ -315,6 +327,87 @@ def test_skip_version_check(
     mock_env = {"PLATFORM_TOOLS_SKIP_VERSION_CHECK": mock_env_var} if mock_env_var else {}
     with patch.dict(os.environ, mock_env):
         assert skip_version_check() == expected
+
+
+class TestPlatformHelperVersioningGetStatus:
+    def test_get_platform_helper_version_status_given_config_and_deprecated_version_file(
+        self,
+        valid_platform_config,
+        mock_pypi_provider,
+        mock_local_version_provider,
+        mock_file_provider,
+        mock_config_provider,
+    ):
+        mock_local_version_provider.get_installed_tool_version.return_value = SemanticVersion(
+            1, 1, 1
+        )
+        mock_pypi_provider.get_latest_version.return_value = SemanticVersion(2, 3, 4)
+
+        mock_file_provider.load.return_value = "5.6.7"
+        mock_config_provider.load_unvalidated_config_file.return_value = valid_platform_config
+
+        version_status = PlatformHelperVersioning(
+            local_version_provider=mock_local_version_provider,
+            pypi_provider=mock_pypi_provider,
+            config_provider=mock_config_provider,
+            file_provider=mock_file_provider,
+        ).get_version_status()
+
+        assert version_status.local == SemanticVersion(1, 1, 1)
+        assert version_status.latest == SemanticVersion(2, 3, 4)
+        assert version_status.deprecated_version_file == SemanticVersion(5, 6, 7)
+        assert version_status.platform_config_default == SemanticVersion(10, 2, 0)
+        assert version_status.pipeline_overrides == {"test": "main", "prod-main": "9.0.9"}
+
+    # TODO clean up mocking
+    @patch("requests.get")
+    def test_get_platform_helper_version_status_with_invalid_yaml_in_platform_config(
+        self, mock_latest_release_request, fakefs
+    ):
+        mock_local_version = Mock()
+        mock_local_version.get_installed_tool_version.return_value = SemanticVersion(1, 1, 1)
+
+        mock_latest_release_request.return_value.json.return_value = {
+            "releases": {"1.2.3": None, "2.3.4": None, "0.1.0": None}
+        }
+        fakefs.create_file(PLATFORM_HELPER_VERSION_FILE, contents="5.6.7")
+        fakefs.create_file(PLATFORM_CONFIG_FILE, contents="{")
+
+        version_status = PlatformHelperVersioning(
+            local_version_provider=mock_local_version
+        ).get_version_status()
+
+        assert version_status.local == SemanticVersion(1, 1, 1)
+        assert version_status.latest == SemanticVersion(2, 3, 4)
+        assert version_status.deprecated_version_file == SemanticVersion(5, 6, 7)
+        assert version_status.platform_config_default == None
+        assert version_status.pipeline_overrides == {}
+
+    @patch("requests.get")
+    def test_get_platform_helper_version_status_with_invalid_config(
+        self,
+        mock_get,
+        fakefs,
+        create_invalid_platform_config_file,
+    ):
+
+        mock_local_version = Mock()
+        mock_local_version.get_installed_tool_version.return_value = SemanticVersion(1, 1, 1)
+
+        mock_get.return_value.json.return_value = {
+            "releases": {"1.2.3": None, "2.3.4": None, "0.1.0": None}
+        }
+        fakefs.create_file(PLATFORM_HELPER_VERSION_FILE, contents="5.6.7")
+
+        version_status = PlatformHelperVersioning(
+            local_version_provider=mock_local_version
+        ).get_version_status()
+
+        assert version_status.local == SemanticVersion(1, 1, 1)
+        assert version_status.latest == SemanticVersion(2, 3, 4)
+        assert version_status.deprecated_version_file == SemanticVersion(5, 6, 7)
+        assert version_status.platform_config_default == SemanticVersion(1, 2, 3)
+        assert version_status.pipeline_overrides == {"prod-main": "9.0.9"}
 
 
 class TestPlatformHelperVersioningCheckIfNeedsUpdate:
